Separate steps into components
